Updates and refactoring

+++ src/st-journal-svc/pkg/models/trade.go
@@ -9,14 +9,14 @@ type Trade struct {
 	Instrument   string    `json:"instrument"`
 	Comments     string    `json:"comments"`
 	Direction    string    `json:"direction"`
-	EntryPrice   uint64    `json:"entryPrice"`
-	ExitPrice    uint64    `json:"exitPrice"`
+	EntryPrice   float32   `json:"entryPrice"`
+	ExitPrice    float32   `json:"exitPrice"`
 	Market       string    `json:"market"`
 	Outcome      string    `json:"outcome"`
-	Quantity     uint32    `json:"quantity"`
-	StopLoss     uint64    `json:"stopLoss"`
+	Quantity     float32   `json:"quantity"`
+	StopLoss     float32   `json:"stopLoss"`
 	Strategy     string    `json:"strategy"`
-	TakeProfit   uint64    `json:"takeProfit"`
+	TakeProfit   float32   `json:"takeProfit"`
 	TimeClosed   time.Time `json:"timeClosed"`
 	TimeExecuted time.Time `json:"timeExecuted"`
 }
